refactor(LinkTreeWidget): Extract LinkTree
